添加条件判断：新文件是否存在

+++ shellCode/publish_bikecc.sh
@@ -2,6 +2,7 @@
 #!/bin/bash
 BIKECC_SDIR=/home/bike/apache-tomcat-bikecc
 StartTomcat=/home/bike/apache-tomcat-bikecc/bin/
+newfile=/home/bikegw/war_new/bikecc.war
 #获取时间并格式化
 time=`date +"%Y%m%d%H%M" `
 #备份文件名
@@ -22,23 +23,25 @@ fi;
 echo "tomcat 关闭成功";
 
 sleep 3
-
-echo "-------------------  开始备份并删除日志文件与旧 war 包  ----------------------";
-#拷贝文件至至war_backup目录，没有则新建目录
-mkdir -p /home/bikecc/war_backup/ && cp ${BIKECC_SDIR}/webapps/bikecc.war /home/bikecc/war_backup/$filename
-echo "-----------------------   备份成功！  ---------------------------"
-rm -rf ${BIKECC_SDIR}/logs/*
-rm -rf ${BIKECC_SDIR}/bin/logs/*
-rm -rf ${BIKECC_SDIR}/webapps/*
-echo "-----------------------   删除成功！  ---------------------------"
-cp /home/bike/war_new/whbike.war ${BIKECC_SDIR}/webapps/
-echo "-----------------------   新 war 拷贝成功！  --------------------------"
-echo "-----------------------   开始启动 bikecc Tomcat  --------------------------"
-#获取环境变量，否则提示找不到 NEST_HOME
-export NEST_HOME=/home/bike/NEST_FILE
-cd $StartTomcat
-nohup ./startup.sh &
-sleep 4 #等候4秒
-TomcatNEWID=$(ps -ef |grep tomcat |grep -w 'apache-tomcat-bikecc'|grep -v 'grep'|awk '{print $2}')
-echo "bikecc tomcat 新进程为:$TomcatNEWID"
-
+if [ ! -f "$newfile" ]; then
+  echo "bikecc.war不存在，请上传文件！"
+else
+   echo "-------------------  开始备份并删除日志文件与旧 war 包  ----------------------";
+   #拷贝文件至至war_backup目录，没有则新建目录
+   mkdir -p /home/bikecc/war_backup/ && cp ${BIKECC_SDIR}/webapps/bikecc.war /home/bikecc/war_backup/$filename
+   echo "-----------------------   备份成功！  ---------------------------"
+   rm -rf ${BIKECC_SDIR}/logs/*
+   rm -rf ${BIKECC_SDIR}/bin/logs/*
+   rm -rf ${BIKECC_SDIR}/webapps/*
+   echo "-----------------------   删除成功！  ---------------------------"
+   cp /home/bike/war_new/whbike.war ${BIKECC_SDIR}/webapps/
+   echo "-----------------------   新 war 拷贝成功！  --------------------------"
+   echo "-----------------------   开始启动 bikecc Tomcat  --------------------------"
+   #获取环境变量，否则提示找不到 NEST_HOME
+   export NEST_HOME=/home/bike/NEST_FILE
+   cd $StartTomcat
+   nohup ./startup.sh &
+   sleep 4 #等候4秒
+   TomcatNEWID=$(ps -ef |grep tomcat |grep -w 'apache-tomcat-bikecc'|grep -v 'grep'|awk '{print $2}')
+   echo "bikecc tomcat 新进程为:$TomcatNEWID"
+fi
